Version 0.25 + Optimizations
Turns have been implemented along with timers. Possible moves are working, only moves to implement are passant, castling and pawn promotion. Pieces cannot move if they place the king under check either. Next steps are to implement the special moves, implement checkmates and stalemates then we can move onto the start and end screen.

Optimizations:
Changed GAME_PIECES variable in board to an array, properly sorted it since the sorting before wasn't working, and made GAME_PIECES a static variable in the App class so it can be used for castling, passant and pawn promotion. The GAME_PIECES variable is also used for checking if the king is placed under check.

TO-DO: I am planning on changing the method inside of get possible moves from queen, bishop and rook so they are iteratively done to preserve memory and speed.