Bug Fixes
Many, MANY things wrong with the earlier check code. Comepletely forgot that a pawn and the other king can place  your own king in check. Also fixed an issue with the diagonal and vertical and horizontal checks since they would still continue to loop after finding a piece that poses no threat to it. It should be breaking out of the loop instead.