Start Screen mostly finished.
Most of the start screen is finished, just need to implement the animations for the settings button and see if I can stop the flickering effect. I also plan on making a ButtonTransition class so this fade transition can be applied to all buttons. Next steps are to implement the animation for settingsButton and to begin working on Settings screen.