Added config file for settings.
I created the config file for the settings and set it up so if it ever gets deleted it remakes one. Also added author declarations to some classes.

Next I need to add action events for each of the menu items, as well as transitions to make everything look nicer.