Added documentation and laid out the bot class.
Did some documentation of methods and classes and also laid out the foundation of the bot class. I also optimized the settings controller and the game controller by using loops to do create the transitions instead of making the transition for each one. A lot less repetition in the code now.