Added the start of online features
Implemented a way for users to register an account, when they load into the home screen after signing in they should be able to see their display name and their avatar.

Next up is to working on the sign in page and also setting up a cloud storage system for the users.