Made some changes to some methods in the bot class.
rookMoves() generates all the moves as if the piece was moving like a rook. I also cleared up the generateBoards method so only code that works is there. I also realized there was  a bug in the diagonal move generation code that I need to fix. In general I fixed many bugs, so many I forget some of them.