A.I. is still getting done.
I made a basic evaluate function, however I ran into some bugs and had to make some changes that I forget but it should work now. However, currently the minimax doesn't work since it returns the wrong value, still needs to be fixed. I also fixed bugs with the passant and castle that allowed a castle to castle with another teams piece and a passant move to place the king in check.