More A.I. development and am in the process of redoing constants.
A.I. is closer to being there, just need to optimize things and improve the evaluation method. Still have  to add alpha beta pruning to the algorithm. One optimization requires me to redo the constants, I am in the process of doing that. I haven't fully done it yet but I am in the process of making new constants.