Optimizations in Bot.
Removed some conditions that needed to be checked since promoted rooks were now properly included in the piece ids, which allows them to be included in the ranges. I also changed the value that the sorting value is based on since last one was incorrect. The bot doesn't take super long to generate moves for a depth of 5 now which is good.